Implement scraping and class for an asset
Finished selenium script that scrapes a page for the listings. Separates them into bid/ask groups as min/max heaps. The class for the asset will take in a driver and link to the asset's page and arbitrage if there is an opportunity. Need to write the actual steam trade execution functionality